Show key popup when one is pressed, and handle key repeat for backspace

+++ src/util/framebuffer.hpp
@@ -4,18 +4,19 @@
 #include <cstddef>
 #include <mutex>
 #include <atomic>
+#include "offset.hpp"
 
 struct Framebuffer {
 	std::mutex mutex{};
 
 	GLuint framebuffer = 0;
 	GLuint texture = 0;
-	size_t width = 0;
-	size_t height = 0;
+	uint32_t width = 0;
+	uint32_t height = 0;
 
-	Framebuffer(size_t width, size_t height);
+	Framebuffer(uint32_t width, uint32_t height);
 
-	void resize(size_t new_width, size_t new_height);
+	void resize(uint32_t new_width, uint32_t new_height);
 
 	~Framebuffer();
 };
